requirements.txt
certifi==2020.6.20
cycler==0.10.0
imageio==2.9.0
kiwisolver==1.2.0
matplotlib==3.3.2
numpy==1.19.2
opencv-python==4.4.0.44
pandas==1.1.3
Pillow==8.0.1
pyparsing==2.4.7
pytesseract==0.3.6
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
